MindMatrix/DataOrganizerAI/requirements.txt
Pillow
tqdm
PyPDF2
moviepy
pydub



